requirements.txt
pyserial>=3.5

vapi-python>=0.1.9
speechrecognition>=3.10.0
pyaudio>=0.2.11
transformers>=4.0.0
torch>=1.0.0
textblob>=0.17.1

gitpython>=3.1.0

fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0

pydantic>=1.8.0
python-dotenv>=0.19.0
pyyaml>=6.0
loguru>=0.6.0

pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910